% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_SpatioTemporal_Rjulia.R
\name{cluster_SpatioTemporal_R}
\alias{cluster_SpatioTemporal_R}
\title{Find spatiotemporal connected clusters in a three-dimensional array}
\usage{
cluster_SpatioTemporal_R(
  arr,
  ncell_connect = 1,
  ncell_overlap = 5,
  factor = 10000,
  diag = FALSE,
  verbose = TRUE,
  version = c("julia", "fortran"),
  .parallel = FALSE,
  nodes = 16,
  ...
)
}
\arguments{
\item{arr}{a three-dimensional array including only TRUE and FALSE.}

\item{ncell_connect}{a integer. If a cluster with number of grids is no more
than this threshold, this cluster will be excluded.}

\item{ncell_overlap}{a integer. If the share grids of the two clusters in two
consecutive layers are no less than the overlap, the two clusters will be
taken as the same one.}

\item{factor}{a integer that is used to recode cluster encoding.}

\item{diag}{a logical value. If TRUE, the diagonal grids are taken as
the adjacent congeneric grids.}

\item{verbose}{a logical value. If TRUE, the cluster translation process will
be printed.}

\item{nodes}{a integer. The number of cluster will be opened for parallel
calculation.}
}
\description{
Connect clusters in space and time
}
