% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterID_recode.R
\name{clusterID_recode}
\alias{clusterID_recode}
\title{Recode encodings of clusters}
\usage{
clusterID_recode(clusterID, factor = 10000)
}
\arguments{
\item{clusterID}{a ngrid x ntime matrix. The clusters in this matrix have
been found by the function connect_spatial_Fortran.}

\item{factor}{a integer that is used to recode cluster encoding.}
}
\value{
a list including clusterID, cno, ncluster, and mask:
\describe{
\item{clusterID}{the same as para clusterID but recoded}
\item{cno}{a ntime x ncluster matrix with cluster encodings}
\item{ncluster}{a vector with the number of clusters in each time}
\item{mask}{a vector to show which grids have clusters}
}
}
\description{
Recode encodings of clusters by a factor
}
\keyword{internal}
