% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterID_refactor.R
\name{clusterID_refactor}
\alias{clusterID_refactor}
\title{clusterID_refactor}
\usage{
clusterID_refactor(clusterID, factor = 10000)
}
\arguments{
\item{clusterId}{3d array}
}
\description{
clusterID_refactor
}
\keyword{internal}
