% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStats.R
\name{ClusterStats}
\alias{ClusterStats}
\title{Calculate drought status}
\usage{
ClusterStats(mat, mask, SMI_thld, idCluster, shortCnoList,
  mGridArea = NULL)
}
\arguments{
\item{mat}{double matrix, with the dim of \code{[ngrid, nmonth]}, (\code{ngrid = nrow*ncol}).
SMI drought index.}

\item{mask}{logical matrix, \code{[nrow, ncol]}.}

\item{SMI_thld}{double number,}

\item{idCluster}{array of cluster num_id, with the dim \code{[nrow, ncol, ntime]}.}

\item{shortCnoList}{vector, cluster num_id}

\item{mGridArea}{matrix of grid area}
}
\description{
Calculate drought status
}
