% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_temporal_Rfortran.R
\name{connect_neighbor}
\alias{connect_neighbor}
\title{Connect clusters in two neighbouring times}
\usage{
connect_neighbor(
  r_cluster,
  ntime,
  ncluster,
  mask,
  ncell_overlap = 5,
  verbose = FALSE
)
}
\arguments{
\item{r_cluster}{a list which is the output of the function clusterID_recode}

\item{ntime}{a integer indicates clusters in this time and before are connected}

\item{ncluster}{a vector with the number of clusters in each time}

\item{mask}{a vector to show which grids have clusters}

\item{ncell_overlap}{a integer. If the share grids of the two clusters in two
consecutive layers are no less than the overlap, the two clusters will be
taken as the same one.}

\item{verbose}{a logical value. If TRUE, the cluster translation process will
be printed.}
}
\value{
a list including clusterID and cno
\describe{
\item{clusterID}{the same as para clusterID but recoded}
\item{cno}{a ntime x ncluster matrix with cluster encodings}
}
}
\description{
Find same clusters in two neighbouring times
}
