% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_spatial.R
\name{connect_spatial_backward}
\alias{connect_spatial_backward}
\title{Backward direction to search cluster}
\usage{
connect_spatial_backward(clusterID, i, j)
}
\arguments{
\item{clusterID}{a nxm matrix indicating cluster serial number. It is the
output of the forward search.}

\item{i}{an integer which indicates the ith row.}

\item{j}{an integer which indicates the jth column.}
}
\value{
A matrix with the serial number for the identified cluster.
}
\description{
Search neighbouring grids backward
}
\details{
This function is used to search clusters from bottom-right to top-left
corner. This search is completed after the forward search. The original codes
are sourced from Fortran language and written by
Samaniego et al., nature climate change, 2018.
}
\keyword{internal}
