% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_spatial.R
\name{connect_spatial_matrix}
\alias{connect_spatial_matrix}
\alias{connect_spatial}
\title{Find all clusters in a matrix}
\usage{
connect_spatial_matrix(
  matrix,
  ncell_connected = 1L,
  diag = FALSE,
  .progress = FALSE
)

connect_spatial(arr, ncell_connected = 1L, diag = FALSE, factor = 10000, ...)
}
\arguments{
\item{matrix}{a nxm matrix. Only TRUE and FALSE are included in this matrix.
TRUE indicates this grid is a congeneric grid, while FALSE indicates not.}

\item{ncell_connected}{a integer, \code{cluster_grids >= ncell_connected}}

\item{diag}{a logical value. If TRUE, the diagonal grids are taken as
the adjacent congeneric grids.}

\item{...}{others to \code{\link[=connect_spatial_matrix]{connect_spatial_matrix()}}}
}
\value{
A matrix with the serial number for the identified cluster.
}
\description{
Find all clusters in space
}
\details{
This function searches each grid in the matrix and identifies all
clusters and assigns a serial number for each cluster. The original codes
are sourced from Fortran language and written by
Samaniego et al., nature climate change, 2018.
}
