% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_spatial.R
\name{connect_spatial_Rfortran}
\alias{connect_spatial_Rfortran}
\title{Find all clusters in a matrix}
\usage{
connect_spatial_Rfortran(matrix, ncell_connect, diag = FALSE)
}
\arguments{
\item{matrix}{a nxm matrix. Only TRUE and FALSE are included in this matrix.
TRUE indicates this grid is a congeneric grid, while FALSE indicates not.}

\item{ncell_connect}{a integer. If a cluster with number of grids is no more
than this threshold, this cluster will be excluded.}

\item{diag}{a logical value. If TRUE, the diagonal grids are taken as
the adjacent congeneric grids.}
}
\value{
A matrix with the serial number for the identified cluster.
}
\description{
Find all clusters in space
}
\details{
This function searches each grid in the matrix and identifies all
clusters and assigns a serial number for each cluster. The original codes
are sourced from Fortran language and written by
Samaniego et al., nature climate change, 2018.
}
