% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_temporal_Rfortran.R,
%   R/connect_temporal_Rjulia.R
\name{connect_temporal_Rfortran}
\alias{connect_temporal_Rfortran}
\alias{connect_temporal_Rjulia}
\title{Connect clusters in all times}
\usage{
connect_temporal_Rfortran(r_cluster, ncell_overlap = 5, verbose = FALSE)

connect_temporal_Rjulia(
  r_cluster,
  ncell_overlap = 5,
  factor = 10000,
  verbose = FALSE
)

connect_temporal_Rjulia(
  r_cluster,
  ncell_overlap = 5,
  factor = 10000,
  verbose = FALSE
)
}
\arguments{
\item{r_cluster}{a list which is the output of the function clusterID_recode}

\item{ncell_overlap}{a integer. If the share grids of the two clusters in two
consecutive layers are no less than the overlap, the two clusters will be
taken as the same one.}

\item{verbose}{a logical value. If TRUE, the cluster translation process will
be printed.}
}
\value{
a a ngrid x ntime matrix in which all clusters are connected in space
and time.
}
\description{
Find the same clusters in the whole period
}
