% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{find_cluster}
\alias{find_cluster}
\title{Search adjacent congeneric grids}
\usage{
find_cluster(matrix, clusterID, i, j, miss.val = -999, diag = FALSE)
}
\arguments{
\item{matrix}{a nxm matrix. Only TRUE and FALSE are included in this matrix.
TRUE indicates this grid is a congeneric grid, while FALSE indicates not.}

\item{clusterID}{a nxm matrix. All the vaues in this matrix are the same, i.e. miss.val.}

\item{i}{an integer which indicates the ith row.}

\item{j}{an integer which indicates the jth column.}

\item{miss.val}{a numeric value which is the initial value of the clusterID.}

\item{diag}{a logical value. If TRUE, the diagonal grids are taken as
the adjacent congeneric grids.}
}
\value{
A matrix with the serial number for the identified cluster.
}
\description{
The adjacent congeneric grids are identified by recursion algorithm
}
\details{
For the \link{i, j} grid, this function is used to find all the adjacent
congeneric grids of this \link{i, j} grid. Then, all the identified grids are taken
as a cluster. For this cluster, a serial number is named in the grids of this cluster.
}
