% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fortran_spatial_cluster.R
\name{find_cluster_forward}
\alias{find_cluster_forward}
\title{Forward direction to search cluster}
\usage{
find_cluster_forward(matrix, clusterID, i, j)
}
\arguments{
\item{matrix}{a nxm matrix. Only TRUE and FALSE are included in this matrix.
TRUE indicates this grid is a congeneric grid, while FALSE indicates not.}

\item{clusterID}{a nxm matrix. All the initial values in this matrix are NA.}

\item{i}{an integer which indicates the ith row.}

\item{j}{an integer which indicates the jth column.}
}
\value{
A matrix with the serial number for the identified cluster.
}
\description{
Search neighbouring grids forward
}
\details{
This function is used to search clusters from top-left to bottom-right
corner. The original codes are sourced from Fortran language and written by
Samaniego et al., nature climate change, 2018.
}
