% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave-package.R
\docType{package}
\name{heatwave-package}
\alias{heatwave}
\alias{heatwave-package}
\title{heatwave: heatwave}
\description{
Heatwaves in spatial.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
