% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{InitCluster}
\alias{InitCluster}
\title{Open multi-clusters for parallel calculation}
\usage{
InitCluster(ncluster = 16, outfile = "log.txt", kill = TRUE)
}
\arguments{
\item{ncluster}{the number of cluster will be opened for parallel calculation}

\item{outfile}{the output file for open cluster information}

\item{kill}{if TRUE the opened clusters will be stopped}
}
\description{
Open multi-clusters for parallel calculation and be suitable for unix and windows
}
