% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cluster}
\alias{plot.cluster}
\title{Plot drought cluster result}
\usage{
\method{plot}{cluster}(r2, x = NULL, y = NULL)
}
\arguments{
\item{r2}{obj returned by [ClusterEvolution()]}

\item{x}{numeric vector, latitudes}

\item{y}{numeric vector, longitudes}
}
\description{
Plot drought cluster result
}
