% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cluster}
\alias{plot.cluster}
\title{Plot drought cluster result}
\usage{
plot.cluster(idClusters, times = 1:4, range = NULL,
  origin = "1961-01-01", ..., sp.layout = NULL)
}
\arguments{
\item{idClusters}{obj returned by \code{\link[=ClusterEvolution]{ClusterEvolution()}}}

\item{times}{integer vector, which time to plot?}

\item{range}{2 numeric vector}

\item{...}{other parameters to \code{\link[lattice:levelplot]{lattice::levelplot()}}}

\item{x}{numeric vector, latitudes}

\item{y}{numeric vector, longitudes}
}
\description{
Plot drought cluster result
}
