% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMI-package.R
\docType{package}
\name{SMI-package}
\alias{SMI}
\alias{SMI-package}
\title{SMI: SMI}
\description{
One paragraph description of what the package does as one
        or more full sentences.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
