% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{spatial_cluster}
\alias{spatial_cluster}
\title{Find all clusters in a matrix}
\usage{
spatial_cluster(matrix, miss.val = -999, diag = FALSE)
}
\arguments{
\item{matrix}{a nxm matrix. Only TRUE and FALSE are included in this matrix.
TRUE indicates this grid is a congeneric grid, while FALSE indicates not.}

\item{miss.val}{a numeric value which is the initial value of the clusterID in
the function find_cluster.}

\item{diag}{a logical value. If TRUE, the diagonal grids are taken as
the adjacent congeneric grids.}
}
\value{
A matrix with serial numbers for all identified clusters.
}
\description{
Find all clusters in space with the function find_cluster
}
\details{
This function searches each grid in the matrix and identifies all
clusters and assigns a serial number for each cluster. This function is built
based on the find_cluster.
}
