% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fortran_spatial_cluster.R
\name{spatial_cluster_Fortran}
\alias{spatial_cluster_Fortran}
\title{Find all clusters in a matrix}
\usage{
spatial_cluster_Fortran(matrix, thres, diag = FALSE)
}
\arguments{
\item{matrix}{a nxm matrix. Only TRUE and FALSE are included in this matrix.
TRUE indicates this grid is a congeneric grid, while FALSE indicates not.}

\item{thres}{a integer. If a cluster with number of grids is no more than this
threshold, this cluster will be excluded.}

\item{a}{logical value. If TRUE, the diagonal grids are taken as
the adjacent congeneric grids.}
}
\value{
A matrix with the serial number for the identified cluster.
}
\description{
Find all clusters in space
}
\details{
This function searches each grid in the matrix and identifies all
clusters and assigns a serial number for each cluster. The original codes
are sourced from Fortran language and written by
Samaniego et al., nature climate change, 2018.
}
