% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_SpatioTemporal_Fortran.R,
%   R/cluster_SpatioTemporal_Rjulia.R
\name{cluster_SpatioTemporal}
\alias{cluster_SpatioTemporal}
\alias{cluster_SpatioTemporal_julia}
\alias{cluster_SpatioTemporal_R}
\title{cluster_SpatioTemporal}
\usage{
cluster_SpatioTemporal(
  arr,
  cellCoor = NULL,
  ncell_connected = 1L,
  ncell_overlap = 1L,
  factor = 10000,
  diag = FALSE
)

cluster_SpatioTemporal_julia(
  arr,
  ncell_connected = 1L,
  ncell_overlap = 5L,
  factor = 10000,
  diag = FALSE,
  ...
)

cluster_SpatioTemporal_R(
  arr,
  ncell_connected = 1L,
  ncell_overlap = 5L,
  factor = 10000,
  diag = FALSE,
  verbose = FALSE,
  version = c("julia", "fortran"),
  ...
)
}
\arguments{
\item{arr}{integer array, \verb{[nrow, ncol, ntime]}, returned by \code{\link[=eventIndicator]{eventIndicator()}}}

\item{cellCoor}{integer matrix, \verb{[nCells, 2]}}

\item{ncell_connected}{treshold  for cluster formation in space}

\item{ncell_overlap}{number cells for joining clusters in time}

\item{factor}{max number of clusters per time}

\item{version}{character, "julia" or fortran}
}
\description{
cluster drought clusters in space and time
}
\examples{
set.seed(1)

nlon <- 10
nlat <- 10
ngrid <- nlon*nlat
ntime <- 12
dim = c(nlon, nlat, ntime)
threshold = 0.2
 
arr <- array(rnorm(nlon*nlat*ntime), dim = c(nlon, nlat, ntime)) > 0.2 # boolean arr
# mask = apply_3d(arr >= 0.2, FUN = rowSums2) > 0
# r  <- droughtIndicator( arr, mask, threshold = 0.2)
r_cluster <- cluster_SpatioTemporal(arr, ncell_connected = 1, ncell_overlap = 1, factor = 1000)
# plot.cluster(r_cluster$clusterId)
# r_status  <- ClusterStats(mat, mask, SMI_thld, r_cluster$clusterId, r_cluster$shortCnoList)
}
\author{
Budapest, 10-11.03.2011
}
