% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterEvolution.R
\name{ClusterEvolution}
\alias{ClusterEvolution}
\title{ClusterEvolution}
\usage{
ClusterEvolution(SMIc, cellCoor, thCellClus = 1L, nCellInter = 1L)
}
\arguments{
\item{SMIc}{integer array, \code{[nrow, ncol, ntime]}}

\item{cellCoor}{integer matrix, \code{[nCells, 2]}}

\item{thCellClus}{treshold  for cluster formation in space}

\item{nCellInter}{number cells for joining clusters in time}
}
\description{
cluster drought clusters in space and time
}
\author{
Budapest, 10-11.03.2011
}
