% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droughtCluster.R
\name{ClusterEvolution}
\alias{ClusterEvolution}
\title{ClusterEvolution}
\usage{
ClusterEvolution(SMIc, cellCoor, nCellInter = 1, thCellClus = 1)
}
\arguments{
\item{SMIc}{integer array, `[nrow, ncol, ntime]`}

\item{cellCoor}{integer matrix, `[nCells, 2]`}

\item{nCellInter}{number cells for joining clusters in time}

\item{thCellClus}{treshold  for cluster formation in space}
}
\description{
cluster drought clusters in space and time
}
\author{
Budapest, 10-11.03.2011
}
