% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterEvolution.R
\name{ClusterEvolution}
\alias{ClusterEvolution}
\title{ClusterEvolution}
\usage{
ClusterEvolution(SMIc, cellCoor, thCellClus = 1L, nCellInter = 1L)
}
\arguments{
\item{SMIc}{integer array, \verb{[nrow, ncol, ntime]}}

\item{cellCoor}{integer matrix, \verb{[nCells, 2]}}

\item{thCellClus}{treshold  for cluster formation in space}

\item{nCellInter}{number cells for joining clusters in time}
}
\description{
cluster drought clusters in space and time
}
\examples{
# devtools::load_all()
library(heatwave)

{
    set.seed(1)
    n <- 100
    nrow <- n
    ncol <- n
    ngrid <- nrow * ncol
    ntime <- 12

    mat <- matrix(rnorm(nrow * ncol * ntime), ngrid, ntime)
    SMI_thld <- 0.5
    mask <- matrix(mat[, 1], nrow, ncol) > SMI_thld
    arr <- array(mat, dim = c(nrow, ncol, ntime)) > SMI_thld
}

{
    library(JuliaCall)
    julia_setup(rebuild = TRUE)
    julia_source("cluster/src/cluster.jl")
    # system.time(res <- julia_call("cluster.spatial_cluster", arr))
    # names(res) <- c("nC", "cno", "IdClusters")
    res = julia_call("cluster.cluster_spatiotemporal", arr)
}

r <- droughtIndicator(mat, mask, SMI_thld)
r_cluster <- ClusterEvolution(r$SMIc, r$cellCoor)

{
    system.time(r_status <- ClusterStats(mat, mask, SMI_thld, r_cluster$idCluster, r_cluster$shortCnoList))
}

## 2. visualization ------------------------------------------------------------
# devtools::install_github("kongdd/Ipaper")
{
    p <- plot.cluster(r_cluster$idCluster, 1:6)
    ratio <- 0.8
    write_fig(p, "b.pdf", 12 * ratio, 8 * ratio)
}

{
    p <- plot.cluster(idC, 1:6)
    ratio <- 0.8
    write_fig(p, "r_julia.pdf", 12 * ratio, 8 * ratio)
}

{
    p <- plot.cluster(idC2, 1:6)
    ratio <- 0.8
    write_fig(p, "r_julia_final.pdf", 12 * ratio, 8 * ratio)
}
}
\author{
Budapest, 10-11.03.2011
}
