% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_temporal_Rfortran.R,
%   R/connect_temporal_Rjulia.R
\name{connect_temporal_Rfortran}
\alias{connect_temporal_Rfortran}
\alias{connect_temporal_Rjulia}
\title{Find spatio-temporal clusters in a array}
\usage{
connect_temporal_Rfortran(
  r_cluster,
  ncell_overlap = 5,
  factor = 10000,
  verbose = FALSE
)

connect_temporal_Rjulia(
  r_cluster,
  ncell_overlap = 5,
  factor = 10000,
  verbose = FALSE
)
}
\arguments{
\item{ncell_overlap}{a integer. If the share grids of the two clusters in two
consecutive layers are no less than the ncell_overlap, the two clusters will be
taken as the same one.}

\item{factor}{a integer which is used to compile a series number for each
cluster.}

\item{clusterID}{a three-dimensional array. This array is formed layer by
layer. Each layer is a matrix that is the clusters identified by the function
spatial_cluster_Fortran.}
}
\value{
A array. The clusters in param clusterID are connected by temporal
overlaps and renamed by a new series number.
}
\description{
Identify the evolution of regional event
}
\details{
The function is used to connect the temporal links of spatial
clusters identified by the function spatial_cluster_Fortran.
The original codes are sourced from Fortran language and written by
Samaniego et al., nature climate change, 2018.
}
