% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droughtCluster.R
\name{droughtIndicatorR}
\alias{droughtIndicatorR}
\title{droughtIndicator}
\usage{
droughtIndicatorR(mat, mask, SMI_thld)
}
\arguments{
\item{mat}{double matrix, with the dim of `[ngrid, nmonth]`, (`ngrid = nrow*ncol`). 
SMI drought index.}

\item{mask}{logical matrix, `[nrow, ncol]`.}

\item{SMI_thld}{double number,}
}
\value{
* SMIc a drought status array with 0 (non-drought), 1 (drought), -9999 (NA).
* cellCoor index of masked region.
}
\description{
Return a drought status array with 0 (non-drought), 1 (drought), -9999 (NA).
}
\examples{
library(SMI)

set.seed(1)

nrow <- 10
ncol <- 10
ngrid <- nrow*ncol
ntime <- 11

mat <- matrix(rnorm(nrow*ncol*ntime), ngrid, ntime)
mask <- matrix(mat[,1], nrow, ncol) > 0.5
SMI_thld <- 0.2

mode(mat) <- "single"

r  <- droughtIndicatorR( mat, mask, SMI_thld)
r2 <- ClusterEvolution(r$SMIc, r$cellCoor)

p <- plot.cluster(r2)
ratio = 0.8
# devtools::install_github("kongdd/Ipaper")
Ipaper::write_fig(p, "a.pdf", 12*ratio, 8*ratio)
## 2. visualization ------------------------------------------------------------
}
