% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cluster}
\alias{plot.cluster}
\title{Plot drought cluster result}
\usage{
\method{plot}{cluster}(idClusters, x = NULL, y = NULL, ...)
}
\arguments{
\item{idClusters}{obj returned by \code{\link[=ClusterEvolution]{ClusterEvolution()}}}

\item{x}{numeric vector, latitudes}

\item{y}{numeric vector, longitudes}

\item{...}{other parameters to \code{\link[lattice:levelplot]{lattice::levelplot()}}}
}
\description{
Plot drought cluster result
}
